resolving all the missing inline keywords on the dd and qd math functions

+++ static/appenv/multifile/integers.cpp
@@ -1,7 +1,8 @@
 
 // integers.cpp: compilation test to check arithmetic type usage in application environments
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017 Stillwater Supercomputing, Inc.
+// SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/number/integer/integer.hpp>
@@ -10,7 +11,7 @@
 
 using Integer = sw::universal::integer<8, uint8_t, sw::universal::IntegerNumberType::IntegerNumber>;
 
-Integer integerPolynomial(const std::vector<float>& coef, const Integer& x) {
+Integer integerPolynomial(const std::vector<int>& coef, const Integer& x) {
 	using namespace sw::universal;
 	if (coef.size() < 2) {
 		std::cerr << "Coefficient set is too small to represent a polynomial\n";
